Merge from keras-team:master

+++ kerastuner/engine/tuner.py
@@ -67,6 +67,7 @@ class Tuner(base_tuner.BaseTuner):
             by this Tuner.
         logger: Optional. Instance of Logger class, used for streaming data
             to Cloud Service for monitoring.
+        tuner_id: Optional. If set, use this value as the id of this Tuner.
         overwrite: Bool, default `False`. If `False`, reloads an existing project
             of the same name if one is found. Otherwise, overwrites the project.
     """
@@ -114,6 +115,8 @@ class Tuner(base_tuner.BaseTuner):
 
         # Save only the last N checkpoints.
         self._save_n_checkpoints = 10
+
+        self.tuner_id = tuner_id or self.tuner_id
 
     def run_trial(self, trial, *fit_args, **fit_kwargs):
         """Evaluates a set of hyperparameter values.
@@ -163,9 +166,12 @@ class Tuner(base_tuner.BaseTuner):
     def save_model(self, trial_id, model, step=0):
         epoch = step
         self._checkpoint_model(model, trial_id, epoch)
-        if epoch > self._save_n_checkpoints:
+        # TODO: save the top epoch checkpoints instead of last ones.
+        epoch_to_delete = epoch - self._save_n_checkpoints
+        best_epoch = self.oracle.get_trial(trial_id).best_step
+        if epoch > self._save_n_checkpoints and epoch_to_delete != best_epoch:
             self._delete_checkpoint(
-                trial_id, epoch - self._save_n_checkpoints)
+                trial_id, epoch_to_delete)
 
     def load_model(self, trial):
         model = self.hypermodel.build(trial.hyperparameters)
